backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
langchain-core==0.3.45
docker==7.1.0
GitPython==3.1.44

